Rename future visits response
