+++ src/main/java/com/redhat/demo/optaplanner/websocket/WebSocketDownstreamConnector.java
@@ -9,11 +9,7 @@ import com.redhat.demo.optaplanner.Mechanic;
 import com.redhat.demo.optaplanner.websocket.domain.JsonLocation;
 import com.redhat.demo.optaplanner.websocket.domain.JsonMachine;
 import com.redhat.demo.optaplanner.websocket.domain.JsonMechanic;
-import com.redhat.demo.optaplanner.websocket.response.AddMechanicResponse;
-import com.redhat.demo.optaplanner.websocket.response.DispatchMechanicResponse;
-import com.redhat.demo.optaplanner.websocket.response.ConnectResponse;
-import com.redhat.demo.optaplanner.websocket.response.RemoveMechanicResponse;
-import com.redhat.demo.optaplanner.websocket.response.UpdateMachineHealthResponse;
+import com.redhat.demo.optaplanner.websocket.response.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
