+++ src/main/java/com/redhat/demo/optaplanner/websocket/response/ResponseType.java
@@ -22,5 +22,5 @@ public enum ResponseType {
     REMOVE_MECHANIC,
     DISPATCH_MECHANIC,
     UPDATE_MACHINE_HEALTHS,
-    FUTURE_VISITS
+    UPDATE_FUTURE_VISITS
 }
